Refactored the Toeplitz code directory to deload mod.rs and hierarchize the directory overall
